# docs/swagger.yaml
definitions:
...
  controllers.commentphoneInput:
    properties:
      comment:
        type: string
      phone_id:
        type: integer
    type: object
  controllers.newsInput:
    properties:
      content:
        type: string
      link_url:
        type: string
      title:
        type: string
    type: object
  controllers.phoneInput:
    properties:
      brand_id:
        type: integer
      type:
        type: string
      year:
        type: integer
    type: object
...
  models.CommentsPhone:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.News:
    properties:
      content:
        type: string
      created_at:
        type: string
      creator_name:
        type: string
      id:
        type: integer
      link_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Phone:
    properties:
      brand_id:
        type: integer
      created_at:
        type: string
      editor_name:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      year:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /comments-phone:
    get:
      description: Get a list of Comment.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CommentsPhone'
            type: array
      summary: Get all Comment.
      tags:
      - Comment Phone
    post:
      description: Creating a new Comment.
      parameters:
      - description: the body to create a new Comment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.commentphoneInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CommentsPhone'
      security:
      - BearerToken: []
      summary: Create New Comment.
      tags:
      - Comment Phone
  /comments-phone/{id}:
    delete:
      description: Delete a Comment by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Comment.
      tags:
      - Comment Phone
...
  /news:
    get:
      description: Get a list of News.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.News'
            type: array
      summary: Get all News.
      tags:
      - News
    post:
      description: Creating a new News.
      parameters:
      - description: the body to create a new phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.newsInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      security:
      - BearerToken: []
      summary: Create New News. (Admin only)
      tags:
      - News
  /news/{id}:
    delete:
      description: Delete a phone by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one News. (Admin only)
      tags:
      - News
    get:
      description: Get a News by id.
      parameters:
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      summary: Get News by Id.
      tags:
      - News
    patch:
      description: Update phone by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.newsInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      security:
      - BearerToken: []
      summary: Update News. (Admin only)
      tags:
      - News
  /news/{id}/comments:
    get:
      description: Get all Specs and Comment by NewsId.
      parameters:
      - description: News id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get Comments by NewsId.
      tags:
      - News
  /phones:
    get:
      description: Get a list of Phones.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Phone'
            type: array
      summary: Get all phones.
      tags:
      - Phone
    post:
      description: Creating a new Phone.
      parameters:
      - description: the body to create a new phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.phoneInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      security:
      - BearerToken: []
      summary: Create New Phone. (Admin only)
      tags:
      - Phone
  /phones/{id}:
    delete:
      description: Delete a phone by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one phone. (Admin only)
      tags:
      - Phone
    get:
      description: Get a Phone by id.
      parameters:
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      summary: Get Phone.
      tags:
      - Phone
    patch:
      description: Update phone by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.phoneInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      security:
      - BearerToken: []
      summary: Update Phone. (Admin only)
      tags:
      - Phone
  /phones/{id}/specs-comments:
    get:
      description: Get all Specs and Comment by PhoneId.
      parameters:
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get Specs & CommentsPhone.
      tags:
      - Phone
...
swagger: "2.0"